Move preference registration into cogs
