+++ src/ilo/colour.py
@@ -7,12 +7,3 @@ def discord_colours(dictionary):
 
 def rgb_tuple(value):
     return tuple(bytes.fromhex(value))
-
-
-def is_colour(value):
-    try:
-        value = rgb_tuple(value)
-        if len(value) == 3:
-            return True
-    except ValueError:
-        return False
