+++ src/ilo/preferences.py
@@ -1,29 +1,22 @@
 import json
 import os
 
-from ilo.colour import is_colour
-from ilo.defines import acro_choices, text
-from ilo.fonts import fonts
-from ilo.jasima import get_languages_for_slash_commands
+from ilo.defines import text
 
 PREFERENCES_PATH = "userdata/preferences.json"
 
 
 class PreferenceHandler:
     def __init__(self):
-        self.templates = [
-            Template("fontsize", 72, validation=fontsize_validation),
-            Template("color", "ffffff", validation=colour_validation),
-            Template("acro", "ku suli", acro_choices),
-            Template("font", "linja sike", {font: font for font in fonts}),
-            Template("language", "en", get_languages_for_slash_commands()),
-        ]
-        self.templates = {template.name: template for template in self.templates}
+        self.templates = {}
         if os.path.exists(PREFERENCES_PATH):
             self.userdata = self.from_json()
         else:
             self.userdata = {}
             self.to_json()
+
+    def register(self, template):
+        self.templates[template.name] = template
 
     def from_json(self):
         with open(PREFERENCES_PATH, encoding="utf-8") as f:
@@ -85,18 +78,6 @@ class Template:
             if value not in self.choices.values():
                 return False
         return True
-
-
-def fontsize_validation(value):
-    if not (value <= 500 and value >= 14):
-        return "Font size is limited to the range from 14 to 500."
-    return True
-
-
-def colour_validation(value):
-    if not is_colour(value):
-        return "The string has to be a valid hexadecimal rgb colour, e.g. `2288ff`."
-    return True
 
 
 preferences = PreferenceHandler()
